backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
boto3
python-multipart
Pillow
bcrypt
requests